Move some files under repo/setting
